requirements.txt
gradio==4.36.1
Pillow==10.3.0
requests==2.32.3
huggingface-hub==0.23.4